+++ crates/core/action-pipeline/Cargo.toml
@@ -31,7 +31,6 @@ console = { workspace = true }
 miette = { workspace = true }
 num_cpus = "1.16.0"
 proto_core = { workspace = true }
-proto_installer = { workspace = true }
 rustc-hash = { workspace = true }
 serde = { workspace = true }
 starbase_styles = { workspace = true }
