+++ crates/core/actions/src/lib.rs
@@ -1,5 +1,7 @@
+mod install_proto;
 mod sync_codeowners;
 mod sync_vcs_hooks;
 
+pub use install_proto::*;
 pub use sync_codeowners::*;
 pub use sync_vcs_hooks::*;
